scraper/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
psycopg2-binary==2.9.9
lxml==4.9.3
fake-useragent==1.4.0
python-dotenv==1.0.0
pandas==2.1.4

